change in the list of the authors
Because of the number of changes and improvements implemented by @pedbrgs, he is now a contributor to the package. @ocfgaldino and @JulianaStrieder were removed from the list of authors since their participation in the package was concluded and the package is being widely refactored (it will be the version 2.0.0) - we are very grateful for their contributions in the first versions of the package.